requirements.txt
cryptography>=36.0.0
python-dotenv>=0.20.0
redis>=4.2.0